Add files via upload
 Menu
2048 #1
A text based implementation of the game 2048 using python. The game asks the user for the highest rank (need not be 2048 always) that needs to be reached in the game. It is a grid or matrix realization. Initially, the grid will contain sparse non-zero numbers that are lesser than the higher rank, we must merge the equal ranked elements or keys of the grid which is possible if the two are adjacent to eachother in a horizontal or vertical manner. This merge will result in an advancement to the higher ranked element ,a basic addition, and each merge will be scored.