Add files via upload
You are presented with 16 suitcases each of which contains an amount ranging from Rs.1 to Rs.1000000. First, you choose one suitcase out of the lot. This suitcase will remain with you till the end and the money in this suitcase is yours unless, you choose to accept a deal.
Once you have chosen a Suitcase, you start opening the remaining fifteen one at a time. After every threee suitcases, you are presented with a deal.